requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.5
Flask-Session==0.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
Werkzeug>=3.0.0
SQLAlchemy==2.0.21
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0